feature(nsis): kill parent process when detached
This commit introduces `--detached` flag for the installer for the
situations where an electron app initiates the update by spawning a
detached installer with `--updated` flag. In such situations killing the
parent process is permissible and desirable because if it hangs - we
won't be able to overwrite the app executable and installation will
silently succeed with only part of the files updated (7zip extraction
doesn't support reporting errors).